Remove theme color and background color, I can't get it to work on Android.
The index.html theme-color seems to be ignored, so no way to make it color-scheme agnostic, and I don't want a middle-gray status bar in either light or dark mode.

Between the splash screen and the first frame, there is a white screen (which I could customize with a FutureBuilder and load the fonts asynchronously, but I couldn't query light or dark mode at this point either).